+++ src/manager.Creeps.ts
@@ -35,7 +35,7 @@ const ACTIVE_ROLE_CONFIG = new Map([
   // ['Gather-B', { body: BODY_CONFIG['WORKER_R3B'], amount: 1, aheadTime: 80, memory: {role:'Gather', node:1} }],  // W35N57 下方矿点采集
 
   // 收集者
-  ['TR', { basename:'',body: BODY_CONFIG['CARRYER_R3'], amount: 3, aheadTime: 80, memory: {r:'运输', e: 0} }],   // W35N57 将各个节点额外的能量搬运到Storage
+  ['TR', { basename:'',body: BODY_CONFIG['CARRYER_R3'], amount: 1, aheadTime: 80, memory: {r:'运输', e: 0} }],   // W35N57 将各个节点额外的能量搬运到Storage
   // ['GA-B', { body: BODY_CONFIG['采集者R4'], amount: 1, aheadTime: 80, memory: {r:'采集', node:1} }],  // W35N57 下方矿点采集
   // ['GA-A', { body: BODY_CONFIG['采集者R4'], amount: 1, aheadTime: 80, memory: {r:'采集', node:0} }],  // W35N57 下方矿点采集
   // ['Gather-B', { body: BODY_CONFIG['采集者R4'], amount: 1, aheadTime: 80, memory: {role:'Gather', node:1} }],  // W35N57 下方矿点采集
