+++ src/utils/Automatic.ts
@@ -11,7 +11,7 @@ const cleanCreepsMemory = function() : void{
 // 全局定期自动事务
 export const Automatic = {
     run: function(): void{
-        if (Game.time % 100 == 0){
+        if (Game.time % 1000 == 0){
             cleanCreepsMemory();
         }
     },
